+++ charts/victoria-metrics-k8s-stack/CHANGELOG.md
@@ -4,6 +4,7 @@
 - set default DNS domain to `cluster.local.`
 - fixed disabling recording rules in `.Values.defaultRules`
 - updated common dependency 0.0.19 -> 0.0.21
+- fixed cluster variable in etcd dashboard
 
 ## 0.28.2